+++ AddInManager/Model/AssemLoader.cs
@@ -31,6 +31,7 @@ public class AssemLoader
     public void CopyGeneratedFilesBack()
     {
         var files = Directory.GetFiles(tempFolder, "*.*", SearchOption.AllDirectories);
+        if(!files.Any()) return;
         foreach (var text in files)
         {
             if (copiedFiles.ContainsKey(text))