+++ AddInManager/Model/DefaultSetting.cs
@@ -9,8 +9,8 @@ namespace RevitAddinManager.Model;
 /// </summary>
 public static class DefaultSetting
 {
+    public static string AppName = "Revit Add-in Manager";
     public static string FileName = "ExternalTool";
-
     public static string FormatExAddin = ".addin";
     public static string FormatDisable = ".disable";
     public static string AdskPath = "Autodesk\\Revit\\Addins";